PostNewtonian

This package computes orbital dynamics of and waveforms from binary black-hole systems, in the post-Newtonian approximation. Currently, there is no support for eccentric systems, but general precessing quasicircular systems are supported.

Integrating inspiral dynamics

PostNewtonian.inspiral — Function

inspiral(M₁, M₂, χ⃗₁, χ⃗₂, Ωᵢ; kwargs...)

Integrate the orbital dynamics of an inspiraling non-eccentric compact binary.

Required arguments

M₁: Initial mass of black hole 1

M₂: Initial mass of black hole 2

χ⃗₁: Initial dimensionless spin of black hole 1, S⃗₁/M₁²

χ⃗₂: Initial dimensionless spin of black hole 2, S⃗₂/M₂²

Ωᵢ: Initial orbital angular velocity

These parameters all describe the "initial" conditions. See below for an explanation of the different meanings of "initial" and "first" in this context. Note that the masses change in time as a result of tidal heating — though the changes are quite small throughout most of the inspiral. The spins change direction due to precession, but also change in magnitude due to tidal heating. Therefore, the values passed here are only precisely as given precisely at the moment of the initial data corresponding to the frequency Ωᵢ.

Keyword arguments

Ω₁=Ωᵢ: First angular velocity in output data (see next section).

Ωₑ=1: Final angular velocity at which to stop ODE integration.

Rᵢ=Rotor(true): Initial orientation of binary.

PNSys=TaylorT1: Currently the only possibility.

PNOrder=7//2: Not actually used currently.

check_up_down_instability=true: Warn if the "up-down instability" (see below) is likely to affect this system.

time_stepper=AutoVern9(Rodas5()): Choice of solver in OrdinaryDiffEq to integrate ODE.

abstol=eps(T)^(11//16): Absolute tolerance of ODE solver, where T is the common type to which all the positional arguments are promoted. This is the tolerance on local error estimates, not necessarily the global error.

reltol=eps(T)^(11//16): Relative tolerance of ODE solver. (As above.)

termination_criteria_forwards=nothing: Callbacks to use for forwards-in-time evolution. See below for discussion of the default value.

termination_criteria_backwards=nothing: Callbacks to use for backwards-in-time evolution. See below for discussion of the default value.

force_dtmin=true: If dt decreases below the integrator's own minimum, and this is false, the integrator will immediately raise an error, before the termination criteria have the chance to exit gracefully. Note that a true value here is critical if the dtmin_terminator callback is to have any effect.

quiet=false: If set to true, informational messages about successful terminations of the ODE integrations (which occur when the target $v$ is reached in either direction) will be silenced. Warnings will still be issued when terminating for other reasons; if you wish to silence them too, you should do something like
using Logging
with_logger(SimpleLogger(Logging.Error)) do
    <your code goes here>
end

integrate_orbital_phase=false: If set to true, integrate the orbital phase $Φ$ along with the rest of the system. Note that this may slow the system down because the absolute value of $Φ$ may grow to very large values, so that the abstol will strain to keep its evolution far more accurate than is really needed. If this is a problem, you can loosen abstol and/or pass vectors of separate tolerances for each variable in the ODE system (see below).

All remaining keyword arguments are passed to the solve function of DiffEqBase. See that function's documentation for details, including useful keyword arguments. The most likely important ones are

saveat: Denotes specific times to save the solution at, during the solving phase.

dt: Sets the initial stepsize. Defaults to an automatic choice if the method is adaptive.

dtmax: Maximum dt for adaptive timestepping.

dtmin: Minimum dt for adaptive timestepping.

Note that if you want the solution to be output with specifically spaced time steps, you don't want dt, which is just the initial suggestion for adaptive systems; you want to set saveat to the desired spacing. [The saveat argument could be a vector of specific times at which to save, but because we don't know when the PN evolution ends, this probably isn't useful.]

Also note that callback can be used, and is combined with the callbacks generated by the termination_criteria_* arguments above. See the documentation for more details, but note that if you want to make your own callbacks, you will need to add OrdinaryDiffEq to your project — or possibly even DifferentialEquations for some of the fancier built-in callbacks.

ODE system

The evolved variables, in order, are

M₁: Mass of black hole 1

M₂: Mass of black hole 2

χ⃗₁ˣ: $x$ component of dimensionless spin of black hole 1

χ⃗₁ʸ: $y$ component...

χ⃗₁ᶻ: $z$ component...

χ⃗₂ˣ: $x$ component of dimensionless spin of black hole 2

χ⃗₂ʸ: $y$ component...

χ⃗₂ᶻ: $z$ component...

Rʷ: Scalar component of frame rotor

Rˣ: $x$ component...

Rʸ: $y$ component...

Rᶻ: $z$ component...

v: PN "velocity" parameter related to the total mass $M$ and orbital angular velocity $Ω$ by $v = (M Ω)^{1/3}$

Φ: Orbital phase given by integrating $Ω$ (optional; only appears if integrate_orbital_phase is true)

The masses and spin magnitudes evolve according to tidal_heating. The spin directions evolve according to Ω⃗ᵪ₁ and Ω⃗ᵪ₂. The frame rotor $R$ is given by integrating the angular velocity as described in Boyle (2016), while the angular velocity itself is given by Ω⃗ₚ. And finally, the PN parameter $v$ evolves according to something like

\[\dot{v} = - \frac{\mathcal{F} + \dot{M}_1 + \dot{M}_2} {\mathcal{E}'}\]

where 𝓕 is the flux of gravitational-wave energy out of the system and 𝓔′ is the derivative of the binding energy with respect to $v$. For "TaylorT1", the right-hand side of this equation is evaluated as given; for "TaylorT4", the right-hand side is first expanded as a Taylor series in $v$ and then truncated at some desired order; for "TaylorT5", the inverse of the right-hand side is expanded as a Taylor series in $v$, truncated at some desired order, and then inverted to obtain an expression in terms of $v$.

Returned solution

The returned quantity is an ODESolution object, which has various features for extracting and interpolating the data. We'll call this object sol.

Note

The solution comes with data at the time points the ODE integrator happened to step to. However, it also comes with dense output (unless you manually turn it off when calling inspiral). This means that you can interpolate the solution to any other set of time points you want simply by calling it as sol(t) for some vector of time points t. The quantity returned by that will have the following features, just like the original solution. Note that if you only want some of the data you can provide the optional keyword argument idxs to specify which of the elements described below you want to interpolate. For example, if you only want to interpolate the values of M₁ and M₂, you can use sol(t, idxs=[1,2]).

The field sol.t is the set of time points at which the solution is given. To access the ith variable at time step j, use sol[i, j].[1] You can also use colons. For example, sol[:, j] is a vector of all the data at time step j, and sol[i, :] is a vector of the ith variable at all times.

Initial frequency vs. first frequency vs. end frequency

Note the distinction between Ωᵢ (with subscript i) and Ω₁ (with subscript 1). The first, Ωᵢ, represents the angular velocity of the initial condition from which the ODE integrator will begin; the second, Ω₁, represents the target angular velocity of the first element of the output data. That is, the ODE integration will run forwards in time from Ωᵢ to the merger, and then come back to Ωᵢ and run backwards in time to Ω₁. The output data will stitch these two together to be one continuous (forwards-in-time) data series.

For example, if you are trying to match to a numerical relativity (NR) simulation, you can read the masses and spins off of the NR data when the system is orbiting at angular velocity Ωᵢ. Integrating the post-Newtonian (PN) solution forwards in time from this point will allow you to compare the PN and NR waveforms. However, you may want to know what the waveform was at earlier times than are present in the NR data. For this, you also have to integrate backwards in time. We parametrise the point to which you integrate backwards with Ω₁. In either case, element 1 of the output solution will have frequency Ω₁ — though by default it is equal to Ωᵢ.

Similarly, the optional argument Ωₑ=1 is the frequency of the end element of the solution — that is Julia's notation for the last element. Note that this is automatically reduced if necessary so that the corresponding PN parameter $v$ is no greater than 1, which may be the case whenever the total mass is greater than 1.

Up-down instability

Be aware that the up-down instability (where the more massive black hole has spin aligned with the orbital angular velocity, and the less massive has spin anti-aligned) can cause systems with nearly zero precession at the initial time to evolve into a highly precessing system either at earlier or later times. This is a real physical result, rather than a numerical issue. If you want to simulate a truly non-precessing system, you should explicitly set the in-place components of spin to precisely 0. By default, we check for this condition, and will issue a warning if it is likely to be encountered for systems with low initial precession. The function used to compute the unstable region is up_down_instability.

Time-stepper algorithms

Tsit5() is a good default choice for time stepper when using Float64 with medium-low tolerance. If stiffness seems to be impacting the results, AutoTsit5(Rosenbrock23()) will automatically switch when stiffness occurs. For tighter tolerances, especially when using Double64s, Vern9() or AutoVern9(Rodas5()) are good choices. For very loose tolerances, as when using Float32s, it might be better to use OwrenZen3().

Termination criteria

The termination criteria are vital to efficiency of the integration and correctness of the solution. The default values for forwards- and backwards-in-time evolution, respectively, are

CallbackSet(
    termination_forwards(v(Ω=Ωₑ, M=M₁+M₂)),
    dtmin_terminator(T),
    nonfinite_terminator()
)

and

CallbackSet(
    termination_backwards(v(Ω=Ω₁, M=M₁+M₂)),
    dtmin_terminator(T),
    nonfinite_terminator()
)

where T is the common float type of the input arguments. If any additional termination criteria are needed, they could be added as additional elements of the CallbackSets. See the callback documentation for details.

Computing mode weights from the inspiral data

PostNewtonian.h! — Function

h!(h, u; ℓmin=0)
h!(h, M₁, M₂, χ⃗₁ˣ, χ⃗₁ʸ, χ⃗₁ᶻ, χ⃗₂ˣ, χ⃗₂ʸ, χ⃗₂ᶻ, Rʷ, Rˣ, Rʸ, Rᶻ, v; ℓmin=0)
mode_weights!(h, u; ℓmin=0)
mode_weights!(h, M₁, M₂, χ⃗₁ˣ, χ⃗₁ʸ, χ⃗₁ᶻ, χ⃗₂ˣ, χ⃗₂ʸ, χ⃗₂ᶻ, Rʷ, Rˣ, Rʸ, Rᶻ, v; ℓmin=0)

Compute mode weights of gravitational waves emitted by pn system, modifying h in place.

These modes are computed in the "co-orbital" frame, in which the larger object lies on the positive $x$ axis, the smaller lies on the negative $x$ axis, and the instantaneous angular velocity is in the positive $z$ direction.

The modes are stored in h in order of increasing $ℓ$ and increasing $m$, with $m$ iterating fastest, all the way up to the highest available mode, $(8,8)$.

Because gravitational waves have spin weight -2, the $(ℓ,m)=(0,0)$, $(1,-1)$, $(1,0)$, and $(1,1)$ modes are always 0. By default, we assume that these modes are nonetheless included in h. If that is not the case, set ℓmin to the smallest $ℓ$ value that should be present in the output data — ℓmin=2 being the most reasonable alternative.

All non-spinning terms are taken from Blanchet (2014). The 1PN spin-orbit term is from Eq. (3.22d) of Kidder (1995). The 1.5PN spin-orbit term is from Eq. (3.22f) of Kidder (1995) and Eq. (F15b) of Will and Wiseman (1996). The 2PN spin-orbit term is from Eq. (4.13) of Buonanno, Faye, Hinderer (2013), while the 2PN spin-spin term is from Eq. (4.15) of that reference.

Detecting up-down instability

PostNewtonian.up_down_instability — Function

up_down_instability(u)
up_down_instability(M₁, M₂, χ⃗₁ˣ, χ⃗₁ʸ, χ⃗₁ᶻ, χ⃗₂ˣ, χ⃗₂ʸ, χ⃗₂ᶻ, Rʷ, Rˣ, Rʸ, Rᶻ, v)

Compute the range of frequencies over which the system is unstable to increasing precession.

The returned value is a pair of dimensionless frequencies giving the lower and upper frequencies between which we can expect instability. If there is no instability expected, the returned pair is just (1,1).

For compact binaries in which the spins are either aligned or anti-aligned with the orbital angular velocity, we do not expect any precession effects — simply by symmetry. However, if the spin of the higher-mass object is aligned with the orbital angular velocity and the spin of the lower-mass object is anti-aligned, the binary is unstable to precession — meaning that any minuscule misalignment can grow rapidly into significant precession. This was first reported by Gerosa et al. (2015), and the range over which the system is unstable is given by Eq. (2) of that reference. We use the lowest-order approximation to convert binary separation to frequency. The result is also "clamped" between 0 and 1, because sometimes the PN approximations involved break down and return unphysical values.

Index of all functions

PostNewtonian.log2

PostNewtonian.log3

PostNewtonian.log3halves

PostNewtonian.log5

PostNewtonian.log5halves

PostNewtonian.ζ3

PostNewtonian.GWFrames.PNWaveform

PostNewtonian.S

PostNewtonian.aₗ

PostNewtonian.combine_solutions

PostNewtonian.dtmin_terminator

PostNewtonian.h!

PostNewtonian.inspiral

PostNewtonian.nonfinite_terminator

PostNewtonian.n̂

PostNewtonian.q

PostNewtonian.termination_backwards

PostNewtonian.termination_forwards

PostNewtonian.tidal_heating

PostNewtonian.up_down_instability

PostNewtonian.v

PostNewtonian.Σ

PostNewtonian.Ω

PostNewtonian.Ω⃗ᵪ

PostNewtonian.Ω⃗ᵪ₁

PostNewtonian.Ω⃗ᵪ₂

PostNewtonian.Ω⃗ₚ

PostNewtonian.γ

PostNewtonian.δ

PostNewtonian.λ̂

PostNewtonian.μ

PostNewtonian.ν

PostNewtonian.χₐ

PostNewtonian.χₛ

PostNewtonian.χ⃗

PostNewtonian.ℓ̂

PostNewtonian.ℳ

PostNewtonian.𝓔

PostNewtonian.𝓔NS

PostNewtonian.𝓔′

PostNewtonian.𝓕

PostNewtonian.𝓕NS

PostNewtonian.𝛡

1Here, the ith variable just refers to which number it has in the list of evolved variables in the ODE system, as described under "ODE system".