Symbolic manipulations

It can be useful to evaluate the post-Newtonian expressions with symbolic arguments. To do so, we just need to create a PNSystem containing a state vector of symbols. All of the variables defined in PostNewtonian.FundamentalVariables and PostNewtonian.DerivedVariables have methods defined automatically to generate symbols instead of values when called with a symbolic PNSystem. In turn, any function modified by the @pn_expression macro should also be able to return a symbolic result, including all functions described in the "PN expressions" section.

For convenience, this package provides symbolic_pnsystem, which has all the fundamental variables stored as Symbolics.jl variables. We can extract the symbols as usual:

julia> v = PostNewtonian.v(symbolic_pnsystem)
v

More generally, we can obtain a symbolic expression for the binding energy as

julia> 𝓔(symbolic_pnsystem)
(-1//2)*M*ν*(1 + ((19//8)*ν - (27//8) - (1//24)*(ν^2))*(v^4) + (ν*(χ₁₂ + 6(χₐₗ^2)) + [...]

In fact, this is how the derivative-of-binding-energy function 𝓔′ is constructed, essentially as

julia> ∂ᵥ = Differential(v);

julia> expand_derivatives(∂ᵥ(𝓔(symbolic_pnsystem)))
(-1//2)*M*ν*(-90*(v^9)*((M₂*Λ₁) / M₁ + (M₁*Λ₂) / M₂)) + [...]

Note that special care is taken to preserve the types of Irrationals and the arguments of certain functions like sqrt and log. Ordinarily, Julia will evaluate these as Float64s; to ensure that they remain symbolic, we have to wrap them in a function that Symbolics.jl will know not to bother expanding: PostNewtonian.hold. While manipulating these expressions symbolically, you'll probably want to leave those hold calls as they are. If you convert the expressions to code, Julia will compile them away easily, so you don't need to do anything more. However, you can remove them using PostNewtonian.unhold if the code is in Expr form.

PostNewtonian.PostNewtonianSymbolicsExt.hold — Function

hold(x)

Delay evaluation of the argument in Symbolics expressions.

This is just a helper function that acts trivially — like the identity function — but also gets registered with Symbolics to avoid evaluation of the argument. For example, we can preserve expressions like π^2, which Julia would normally convert directly to a Float64.

Note that you probably don't want to use this function directly; this will probably be done for you by @pn_expression or similar. If you do want to use this directly, you probably want another layer of indirection to construct something like Symbolics.Num(SymbolicUtils.Term(hold, [x])) so that you can use the result in a symbolic expression.

PostNewtonian.PostNewtonianSymbolicsExt.unhold — Function

unhold(expr)

Remove occurrences of hold from an Expr.

PostNewtonian.PostNewtonianSymbolicsExt.SymbolicPNSystem — Type

SymbolicPNSystem{T, PNOrder}(state, Λ₁, Λ₂)

A PNSystem that contains information as variables from Symbolics.jl.

See also symbolic_pnsystem for a particular general instance of this type.

PostNewtonian.PostNewtonianSymbolicsExt.symbolic_pnsystem — Constant

symbolic_pnsystem

A symbolic PNSystem that contains symbolic information for all types of PNSystems.

In particular, note that this object has an (essentially) infinite PNOrder, uses the TaylorT1 approximant, and has nonzero values for quantities like Λ₁ and Λ₂. If you want different choices, you may need to call SymbolicPNSystem yourself, or even construct a different specialized subtype of PNSystem (it's not hard).

Examples

julia> using PostNewtonian: M₁, M₂, χ⃗₁, χ⃗₂

julia> M₁(symbolic_pnsystem), M₂(symbolic_pnsystem)
(M₁, M₂)

julia> χ⃗₁(symbolic_pnsystem)
χ⃗₁

julia> χ⃗₂(symbolic_pnsystem)
χ⃗₂