Derived variables

Orbital elements

PostNewtonian.DerivedVariables.n̂ — Function

n̂(pnsystem)
n̂(R)
n_hat(pnsystem)
n_hat(R)

The unit vector pointing from object 2 to object 1, when the frame is given by the rotor R. This is equal to

\[n̂(R) = R x̂ R̄\]

PostNewtonian.DerivedVariables.λ̂ — Function

λ̂(pnsystem)
λ̂(R)
lambda_hat(pnsystem)
lambda_hat(R)

The unit vector pointing in the direction of the instantaneous velocity of object 1, when the frame is given by the rotor R. This is equal to

\[λ̂(R) = R ŷ R̄\]

This also completes the right-handed triple of $(n̂, λ̂, ℓ̂)$.

PostNewtonian.DerivedVariables.ℓ̂ — Function

ℓ̂(pnsystem)
ℓ̂(R)
ell_hat(pnsystem)
ell_hat(R)

The unit vector pointing along the direction of orbital angular velocity, when the frame is given by the rotor R. This is equal to

\[ℓ̂(R) = R ẑ R̄\]

PostNewtonian.DerivedVariables.Ω — Function

Ω(pnsystem)
Ω(;v, M=1)
Omega(pnsystem)
Omega(;v, M=1)

Orbital angular frequency.

The parameter v is the PN velocity parameter, and must be passed as a keyword argument — as in Ω(v=0.1). The parameter M is the total mass of the binary. They are related by definition as

\[\Omega \colonequals \frac{v^3}{M}.\]

See also v.

Mass combinations

PostNewtonian.DerivedVariables.M — Function

M(pnsystem)
M(M₁, M₂)
total_mass(pnsystem)
total_mass(M1, M2)

Compute the total mass $M₁+M₂$.

PostNewtonian.DerivedVariables.μ — Function

μ(pnsystem)
μ(M₁, M₂)
reduced_mass(pnsystem)
reduced_mass(M1, M2)

Compute the reduced mass $(M₁ M₂)/(M₁+M₂)$.

PostNewtonian.DerivedVariables.ν — Function

ν(pnsystem)
ν(M₁, M₂)
reduced_mass_ratio(pnsystem)
reduced_mass_ratio(M1, M2)

Compute the reduced mass ratio $(M₁ M₂)/(M₁+M₂)^2$.

Note that the denominator is squared, unlike in the reduced mass μ.

PostNewtonian.DerivedVariables.δ — Function

δ(pnsystem)
δ(M₁, M₂)
mass_difference_ratio(pnsystem)
mass_difference_ratio(M1, M2)

Compute mass-difference ratio $(M₁-M₂)/(M₁+M₂)$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $δ$ always be positive (or always negative), you are responsible for ensuring that.

PostNewtonian.DerivedVariables.q — Function

q(pnsystem)
q(M₁, M₂)
mass_ratio(pnsystem)
mass_ratio(M1, M2)

Compute mass ratio $M₁/M₂$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $q$ always be greater than or equal to 1 (or vice versa), you are responsible for ensuring that.

PostNewtonian.DerivedVariables.ℳ — Function

ℳ(pnsystem)
ℳ(M₁, M₂)
chirp_mass(pnsystem)
chirp_mass(M1, M2)

Compute the chirp mass ℳ, which determines the leading-order orbital evolution of a binary system due to energy loss by gravitational-wave emission.

The chirp mass is defined as

\[ \mathcal{M} = \frac{(M_1 M_2)^{3/5}} {(M_1 + M_2)^{1/5}}.\]

Spin combinations

PostNewtonian.DerivedVariables.S⃗₁ — Function

S⃗₁(pnsystem)

Dimensionful spin vector of object 1.

PostNewtonian.DerivedVariables.S⃗₂ — Function

S⃗₂(pnsystem)

Dimensionful spin vector of object 2.

PostNewtonian.DerivedVariables.S⃗ — Function

S⃗(pnsystem)
S⃗(M₁, M₂, χ⃗₁, χ⃗₂)

Total (dimensionful) spin vector $S⃗₁+S⃗₂$.

PostNewtonian.DerivedVariables.Σ⃗ — Function

Σ⃗(pnsystem)
Σ⃗(M₁, M₂, χ⃗₁, χ⃗₂)

Differential spin vector $M(a⃗₂-a⃗₁)$.

PostNewtonian.DerivedVariables.χ⃗ — Function

χ⃗(pnsystem)
χ⃗(S⃗, M)

Normalized spin vector $S⃗/M²$.

PostNewtonian.DerivedVariables.χ⃗ₛ — Function

χ⃗ₛ(M₁, M₂, χ⃗₁, χ⃗₂)

Symmetric spin vector $(χ⃗₁+χ⃗₂)/2$.

PostNewtonian.DerivedVariables.χ⃗ₐ — Function

χ⃗ₐ(M₁, M₂, χ⃗₁, χ⃗₂)

Antisymmetric spin vector $(χ⃗₁-χ⃗₂)/2$.

PostNewtonian.DerivedVariables.χₑ — Function

χₑ(s)
chi_eff(s)

Effective spin parameter of the system.

Defined as

\[\chi_{\mathrm{eff}} \colonequals \frac{c}{G} \left( \frac{\mathbf{S}_1}{M_1} + \frac{\mathbf{S}_2}{M_2} \right) \cdot \frac{\hat{\mathbf{L}}_{\mathrm{N}}} {M}.\]

PostNewtonian.DerivedVariables.χₚ — Function

χₚ(s)
chi_p(s)

Effective precession spin parameter of the system.

Note that there are two different definitions of this quantity found in the literature. The original definition (converted to the convention where $M_1 \geq M_2$) is

\[\begin{gathered} A_1 = 2 + \frac{3M_2}{2M_1} \\ A_2 = 2 + \frac{3M_1}{2M_2} \\ \chi_{\mathrm{p}} \colonequals \frac{1}{A_1 M_1^2} \mathrm{max}\left(A_1 S_{1\perp}, A_2 S_{2\perp} \right). \end{gathered}\]

In a paper from early in the detection era, the LIGO collaboration used this definition.

However, a more recent paper redefines this essentially as $M_1^2$ times that quantity. Using the convention that $q = M_2/M_1 \leq 1$, the definition may be more compactly written as

\[\chi_{\mathrm{p}} \colonequals \mathrm{max} \left( \chi_{1\perp}, \chi_{2\perp} q \frac{4q+3}{4+3q} \right).\]

Again, a more recent paper by LIGO/Virgo/KAGRA uses this convention.

Because it seems to be the trend, this function uses the latter definition.

Additionally, there are numerous convenience functions to give certain components of the spins. They take a single pnsystem argument and are not exported. Given the definitions above, they are all fairly self explanatory — such as χ₁², which gives χ⃗₁ ⋅ χ⃗₁; or χ₁₂ = χ⃗₁ ⋅ χ⃗₂; or Sₙ = S⃗ ⋅ n̂. Like all the other fundamental and derived variables, these can be used directly in PN expressions modified by the @pn_expression macro.