Fundamental variables

These constitute the basic quantities needed to fully describe a PN binary system at any instant. All other variables — and ultimately all PN expressions — may be expressed as functions of these. In the docstrings below, pnsystem is an object of any type that subtypes PNSystem.

PostNewtonian.FundamentalVariables.M₁ — Function

M₁(pnsystem)
M1(pnsystem)

Mass of object 1 in this system.

PostNewtonian.FundamentalVariables.M₂ — Function

M₂(pnsystem)
M2(pnsystem)

Mass of object 2 in this system.

PostNewtonian.FundamentalVariables.χ⃗₁ — Function

χ⃗₁(pnsystem)
chi1(pnsystem)

Dimensionless spin vector of object 1 in this system, as a QuatVec.

PostNewtonian.FundamentalVariables.χ⃗₂ — Function

χ⃗₂(pnsystem)
chi2(pnsystem)

Dimensionless spin vector of object 2 in this system, as a QuatVec.

PostNewtonian.FundamentalVariables.R — Function

R(pnsystem)

Orientation of the binary, as a Rotor. This Rotor will rotate the ẑ vector to be along the orbital angular velocity vector and the x̂ vector to be along the separation vector pointing from object 2 to object 1.

See also n̂, λ̂, ℓ̂.

PostNewtonian.FundamentalVariables.v — Function

v(pnsystem)
v(;Ω, M=1)

Post-Newtonian velocity parameter. This is related to the orbital angular frequency $\Omega$ as

\[v \colonequals (M\,\Omega)^{1/3},\]

where $M$ is the total mass of the binary.

Note that if you want to pass the value $Ω$ (rather than a PNSystem), you must pass it as a keyword argument — as in v(Ω=0.1).

See also Ω.

PostNewtonian.FundamentalVariables.Φ — Function

Φ(pnsystem)
Phi(pnsystem)

Integrated orbital phase of the system. It is computed as the integral of Ω.

PostNewtonian.FundamentalVariables.Λ₁ — Function

Λ₁(pnsystem)
Lambda1(pnsystem)

Quadrupolar tidal-coupling parameter of object 1 in this system.

We imagine object 1 begin placed in an (adiabatic) external field with Newtonian potential $\phi$, resulting in a tidal field measured by $\partial_i \partial_j \phi$ evaluated at the center of mass of the object. This induces a quadrupole moment $Q_{ij}$ in object 1, which can be related to the tidal field as

\[Q_{ij} = -\frac{G^4}{c^{10}} \Lambda_1 M_1^5 \partial_i \partial_j \phi,\]

where $M_1$ is the mass of object 1. This tidal-coupling parameter $\Lambda_1$ can be related to the Love number $k_2$ (where the subscript 2 refers to the fact that this is for the $\ell=2$ quadrupole, rather than object 2) as

\[\Lambda_1 = \frac{2}{3} \frac{c^{10}}{G^5} \frac{R_1^5}{M_1^5} k_2,\]

where $R_1$ is the radius of object 1. Note that $\Lambda_1$ is dimensionless. For black holes, it is precisely zero; for neutron stars it may range up to 1; more exotic objects may have significantly larger values.

Note that — as of this writing — only NSNS systems can have a nonzero value for this quantity. All other types return 0, which Julia can use to eliminate code that would then be 0. Thus, it is safe and efficient to use this quantity in any PN expression that specializes on the type of pnsystem.

PostNewtonian.FundamentalVariables.Λ₂ — Function

Λ₂(pnsystem)
Lambda2(pnsystem)

Quadrupolar tidal coupling parameter of object 2 in this system.

See Λ₁ for details about the definition, swapping "object 1" with "object 2".

Note that — as of this writing — only BHNS and NSNS systems can have a nonzero value for this quantity. All other types return 0, which Julia can use to eliminate code that would then be 0. Thus, it is safe and efficient to use this quantity in any PN expression that specializes on the type of pnsystem.