Utilities

Macros

Some of the most useful features of this package are the macros allowing us to write PN expressions in fairly natural form, without worrying about calculating all the variables needed for each expression, or manually accounting for the various PN orders to which we may need to truncate PN expansions. To achieve this, we rely primarily on two macros.

PostNewtonian.extract_var_factor — Method

extract_var_factor(term, var)

Extract a factor of var from the product term.

This is a helper function for var_collect.

PostNewtonian.hold — Method

hold(x)

Delay evaluation of the argument in Symbolics expressions.

This is just a helper function that acts trivially — like the identity function — but also gets registered with Symbolics to avoid evaluation of the argument. For example, we can preserve expressions like π^2, which Julia would normally convert directly to a Float64.

Note that you probably don't want to use this function directly; this will probably be done for you by @pn_expression or similar. If you do want to use this directly, you probably want another layer of indirection to construct something like Symbolics.Num(SymbolicUtils.Term(hold, [x])) so that you can use the result in a symbolic expression.

PostNewtonian.type_converter — Method

type_converter(pnsystem, x)

Convert x to a type appropriate for the float type of pnsystem.

PostNewtonian.unhold — Method

unhold(expr)

Remove occurrences of hold from an Expr.

PostNewtonian.var_collect — Method

var_collect(expr, var)

Collect coefficients in expr of various powers of the variable var.

The inputs should be an Expr and a single Symbol to be found in that Expr.

The return value is an Int representing the highest power of v in the expression, and an Expr representing a Tuple of values corresponding to the coefficients of var to various powers. For example,

julia> PostNewtonian.var_collect(:(1 + a*v + b*v^2 + c*v^4), :v)
4, :((1, a, b, 0, c))

(Note that there was no factor in v^3.) This result is convenient for passing to evalpoly, for example.

PostNewtonian.@pn_expansion — Macro

@pn_expansion [[offset] pnsystem] expansion

Gather terms in expansion by the powers of v involved, the choose on the powers chosen by the pnsystem's PNOrder parameter, and evaluate efficiently in Horner form.

Note that the pnsystem argument can be inserted automatically by @pn_expression. For simplicity of presentation, we will assume that this is done in the examples below.

A "PN expansion" is a polynomial in $v$ for which $ln(v)$ factors may be present in coefficients of $v^k$ for $k≥1$. The input may involve multiple terms with the same power of $v$. (I.e., the expansion does not have to be simplified or collected in sympy parlance.)

However, note that the sum must appear together — as part of the same :+ expression — rather than, say, factored into two sums that multiply each other.

The offset argument represents the index of the relative PN order (that is, 2 times the PN order) at which the PN expansion begins. So if the first term in the expansion is really a relative 2.5-pN term, we should use an offset of 5. For example, the tidal_heating expressions for Ṁ₁ and Ṁ₂ begin with a term proportional to v^15, and are added to the flux 𝓕, which begin's with a term proportional to v^10. This means that Ṁ₁ and Ṁ₂ enter at 2.5-pN relative order, so we use an offset of 5.

PostNewtonian.@pn_expression — Macro

@pn_expression [arg_index=1] func

This macro takes the function func, looks for various symbols inside that function, and if present defines them appropriately inside that function.

The first argument to this macro is arg_index, which just tells us which argument to the function func is a PNSystem. For example, the variables defined in PostNewtonian.FundamentalVariables all take a single argument of pnsystem, which is used to compute the values for those variables; this macro just needs to know where to find pnsystem.

Once it has this information, there are four types of transformations it will make:

For every fundamental or derived variable, the name of that variable used in the body of func will be replaced by its value when called with pnsystem. For example, you can simply use the symbols M₁ or μ in your code, rather than calling them as M₁(pnsystem) or μ(pnsystem) every time they appear.

Every Irrational defined in Base.MathConstants or PostNewtonian.MathConstants will be transformed to the eltype of pnsystem. This lets you naturally use such constants in expressions like 2π/3 without automatically converting to Float64.

Each of a short list of functions given by unary_funcs in utilities/macros.jl will first convert their arguments to the eltype of pnsystem. In particular, you can use expressions like √10 or ln(2) without the result being converted to a Float64.

Insert the pnsystem argument as the first argument to each occurrence of @pn_expansion that needs it.

To be more explicit, the first three are achieved by defining the relevant quantities in a let block placed around the body of func, so that the values may be used efficiently without recomputation.

If you need to use one of the fundamental- or derived-variable functions as arguments of values other than those encapsulated in pnsystem, you'll need to address them explicitly with the module name — as in PostNewtonian.v(;Ω, M).

Manipulating ODE solutions

PostNewtonian.combine_solutions — Method

combine_solutions(sol₋, sol₊)

Combine ODESolutions

This function is internal to this package. It is not entirely general, but allows us to combine the backwards- and forwards-in-time solutions of the PN orbital-evolution ODE equations into a single ODESolution object that should behave just as if it were the result of solve. In particular, indexing, interpolation, and iterations should behave exactly as described in the DifferentialEquations docs.

Termination criteria

Hopefully, it should not be necessary to directly use these termination criteria. They are used by default in the orbital_evolution function. But certain particularly extreme physical parameters may lead ODEs that are difficult to integrate — especially if new PN systems or terms are introduced. These, or similar functions may be helpful examples of "callbacks" that can be passed to the ODE integrator.

Note that several of these will issue warnings if the evolution has to be terminated for particularly bad or suspicious reasons, even if the quiet flag is set to true. See the documentation of the quiet argument to the orbital_evolution function for an example of how to use Logging to quiet even the warnings.

PostNewtonian.decreasing_v_terminator — Function

decreasing_v_terminator([quiet])

Construct termination criterion to stop integration when v is decreasing.

Note that some systems may truly have decreasing v as physical solutions — including eccentric systems and possibly precessing systems. You may prefer to implement another solution, like detecting when v decreases below some threshold, or detecting when v is decreasing too quickly. See this function's source code for a simple

If this terminator is triggered while v is less than 1/2, a warning will always be issued; otherwise an info message will be issued only if the quiet flag is set to false.

PostNewtonian.dtmin_terminator — Function

dtmin_terminator(T, [quiet])

Construct termination criterion to terminate when dt drops below √eps(T).

Pass force_dtmin=true to solve when using this callback. Otherwise, the time-step size may decrease too much within a single time step, so that the integrator itself will quit before reaching this callback, leading to a less graceful exit.

If this terminator is triggered while v is less than 1/2, a warning will always be issued; otherwise an info message will be issued only if the quiet flag is set to false.

PostNewtonian.nonfinite_terminator — Method

nonfinite_terminator()

Construct termination criterion to stop integration when any NaN or Inf is found in the data after an integration step.

If this terminator is triggered, a warning will always be issued.

PostNewtonian.termination_backwards — Function

termination_backwards(v₁, [quiet])

Construct termination criteria of solving PN evolution backwards in time

These criteria include checking that the masses are positive and the dimensionless spins are less than 1, as well as ensuring that the evolution will terminate at v₁.

The optional quiet argument will silence informational messages about reaching the target value of v₁ if set to true, but warnings will still be issued when terminating for other reasons.

PostNewtonian.termination_forwards — Function

termination_forwards(vₑ, [quiet])

Construct termination criteria of solving PN evolution forwards in time

These criteria include checking that the masses are positive and the dimensionless spins are less than 1, as well as ensuring that the evolution will terminate at vₑ.

The optional quiet argument will silence informational messages about reaching the target value of vₑ if set to true, but warnings will still be issued when terminating for other reasons.

Irrational constants

These quantities are constants that appear in PN expressions, so they are not exported, but can be used by importing them explicitly or by using the fully qualified names. They are defined here as Irrationals. This means that Julia can convert them to float types as necessary. Unfortunately, by default Julia converts to Float64. For example, BigFloat(2ζ3) will be a BigFloat, but will only have the precision of a Float64, because 2ζ3 is converted first. To get full precision, you'll need to do things like 2BigFloat(ζ3).

One approach to avoiding this is to explicitly redefine these constants as floats of the desired precision, using let to essentially overwrite the name:

function foo(x)
    let ζ3=oftype(x, ζ3)
        2ζ3 + x
    end
end

Inside the let block, ζ3 is no longer an Irrational; it has been converted to whatever number type x is. Thus, when multiplying by 2, it is not converted to a Float64; its precision matches that of x.

This can be quite awkward, so the macro PostNewtonian.@pn_expression is provided to (among other things) automatically search for all Irrationals and replace them with the appropriate float values.

PostNewtonian.MathConstants.γₑ — Constant

γₑ

Euler's constant (also known as the Euler–Mascheroni constant) is defined as the limit as $n \to \infty$ of the difference between the $n$th partial sum of the harmonic series and $\log(n)$. This is OEIS sequence A001620.

julia> PostNewtonian.γₑ
γₑ = 0.5772156649015...

julia> n=10_000_000; sum(1 ./ (1:n))-log(n)
0.5772157149015307

PostNewtonian.MathConstants.ζ3 — Constant

ζ3
apery

Apéry's constant is defined as $ζ(3)$, where $ζ$ is the Riemann zeta function. This is OEIS sequence A002117.

julia> PostNewtonian.apery
ζ3 = 1.2020569031595...

julia> PostNewtonian.ζ3
ζ3 = 1.2020569031595...

julia> sum((1:10_000_000).^-3)
1.2020569031595896

PostNewtonian.MathConstants.ln2 — Constant

ln2
log2

The natural logarithm of 2. This is OEIS sequence A002162.

julia> PostNewtonian.ln2
ln2 = 0.6931471805599...

julia> exp(PostNewtonian.ln2)
2.0

julia> exp(big(PostNewtonian.ln2))
2.0

PostNewtonian.MathConstants.ln3 — Constant

ln3
log3

The natural logarithm of 3. This is OEIS sequence A002391.

julia> PostNewtonian.ln3
ln3 = 1.0986122886681...

julia> exp(PostNewtonian.ln3)
3.0000000000000004

julia> exp(big(PostNewtonian.ln3))
3.0

PostNewtonian.MathConstants.ln5 — Constant

ln5
log5

The natural logarithm of 5. This is OEIS sequence A016628.

julia> PostNewtonian.ln5
ln5 = 1.6094379124341...

julia> exp(PostNewtonian.ln5)
4.999999999999999

julia> exp(big(PostNewtonian.ln5))
5.0

PostNewtonian.MathConstants.ln³╱₂ — Constant

ln³╱₂
log³╱₂
log3halves

The natural logarithm of 3//2. This is OEIS sequence A016578.

julia> PostNewtonian.ln³╱₂
ln³╱₂ = 0.4054651081081...

julia> exp(PostNewtonian.ln³╱₂)
1.5

julia> exp(big(PostNewtonian.ln³╱₂))
1.5

PostNewtonian.MathConstants.ln⁵╱₂ — Constant

ln⁵╱₂
log⁵╱₂
log5halves

The natural logarithm of 5//2. This is OEIS sequence A016579.

julia> PostNewtonian.ln⁵╱₂
ln⁵╱₂ = 0.9162907318741...

julia> exp(PostNewtonian.ln⁵╱₂)
2.5

julia> exp(big(PostNewtonian.ln⁵╱₂))
2.5

Truncated series

We also have some utilities for dealing with series — or more precisely summations, since we only handle finitely many terms. In particular, we need truncated multiplication (and some times division) of truncated series. This multiplication is associative, and there is an identity element, which means that this structure naturally forms a monoid.

Here, we are assuming that there is a fixed order at which the series are truncated, and that truncated multiplication preserves that order. That is, if $A$ and $B$ are summations in terms up to $v^N$ for some integer $N$, we want their product (and ratio, if it exists) to also be a summation in terms up to $v^N$.

Note that we are not referring to these summations as "polynomials" in $v$, because the coefficients will sometimes involve $\ln(v)$ — which is not technically permitted for polynomials. In particular, the presence of logarithms is irrelevant to our meaning of the "order" of the truncated series. This is standard practice in post-Newtonian theory.[1]

The following functions implement this behavior:

PostNewtonian.truncated_series_inverse — Function

truncated_series_inverse(a)
truncated_series_inverse!(b, a)

Given the coefficients a of a series, find the coefficients b of the multiplicative inverse of that series, up to the order of the original series.

Note that this function returns the coefficients of the inverse, rather than its value. This is relevant for use in truncated_series_product and truncated_series_ratio — the latter of which just combines the former with this function.

For example, suppose the input coefficients represent the series

\[A \colonequals \sum_{i=0}^n a_{i-1} v^i.\]

(Remember that Julia's indexing is 1-based, so we subtract 1 to get the index of the coefficient of $v^i$.) Then we return the coefficients b of the series

\[B \colonequals \sum_{i=0}^n b_{i-1} v^i\]

such that

\[A\, B = 1 + \mathcal{O}(v^{n+1}).\]

Note

This function requires that a[1] be nonzero. If you have a series that starts at a higher term — say, $v^n$ for $n>0$ — you should factor out the $v^n$, and multiply the series resulting from this function by $v^{-n}$.

Explanation

The inverse coefficients can be computed fairly easily by induction. Start by defining

\[b_0 = 1/a_0.\]

Now, assuming that we've computed all coefficients up to and including $b_{i}$, we can compute

\[b_{i+1} = -b_0\sum_{j=1}^{i} a_j b_{i-j}.\]

PostNewtonian.truncated_series_product — Function

truncated_series_product(a, b, v)

Evaluate the truncated product of the series a and b, which are expanded in powers of v.

Note that this function returns the value of the summation, rather than its coefficients.

Here we define the series in terms of the coefficients a and b as

\[A \colonequals \sum_{i=0}^n a_{i-1} v^i \qquad B \colonequals \sum_{i=0}^n b_{i-1} v^i,\]

and return the value of the product $A\, B$ truncated at $v^n$.

Internally, the sums are performed using evalpoly.

See also truncated_series_ratio.

PostNewtonian.truncated_series_ratio — Function

truncated_series_ratio(a, b, v)

Evaluate the truncated ratio of the series a and b, which are expanded in powers of v.

Note that this function returns the value of the summation, rather than its coefficients.

Here we define the series in terms of the coefficients a and b as

\[A \colonequals \sum_{i=0}^n a_{i-1} v^i \qquad B \colonequals \sum_{i=0}^n b_{i-1} v^i,\]

and return the value of the ratio $A / B$ truncated at $v^n$.

This function simply combines truncated_series_product and truncated_series_inverse.

Miscellaneous

PostNewtonian.apply_to_first_add! — Method

apply_to_first_add!(expr, func)

Apply func to the first sub-expression found in a "prewalk"-traversal of expr that satisfies isadd. If func acts in place, so does this function. In either case, the expression should be returned.

PostNewtonian.find_symbols_of_type — Method

find_symbols_of_type(mod, T)

Given a module mod (not just its name, but the actual imported module), find all objects inside that module that are instances of the given type T. The returned quantity is a vector of Symbols.

PostNewtonian.flatten_binary! — Method

flatten_binary!(expr, symbols)

Flatten nested binary operations — that is, apply associativity repeatedly.

PostNewtonian.isadd — Method

isadd(x)

Return true if the Expr x is a call to (+) or :+.

PostNewtonian.iscall — Method

iscall(x, symbols)

Return true if the Expr x is a call to any element of symbols.

PostNewtonian.ismul — Method

ismul(x)

Return true if the Expr x is a call to (*) or :*.

PostNewtonian.value — Method

value(x)

Return x or the value wrapped by the Dual number x

1Different texts in post-Newtonian theory treat these logarithmic terms with varying levels of rigor. The preferred method is Hadamard regularization (often referred to in the literature as the partie finie). A good summary is found in Section 6 of Blanchet's Living Review. Another potential approach could be taken following this paper. But for our purposes, it will suffice to take the simplistic approach of treating logarithmic terms as if they were any other constant.