Computing the waveform

Once you have used orbital_evolution to compute the inspiral, and selected the time steps on which you want the waveform, you can compute the waveform with one of the following functions.

Danger

Remember that, by default, orbital_evolution will return the solution at whatever time steps the ODE integrator chose, which are almost certainly too far apart to satisfy the Nyquist limit relevant for waveforms. You can use the saves_per_orbit argument to that function with a value of at least 2ℓₘₐₓ (where ℓₘₐₓ is the argument to one of the functions below) to barely satisfy the Nyquist limit. I usually use 4ℓₘₐₓ for a little extra safety factor.

PostNewtonian.coorbital_waveform — Function

coorbital_waveform(inspiral; [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Evaluate the post-Newtonian waveform mode weights in the co-orbital frame for the given inspiral output by orbital_evolution.

The PNOrder defaults to the one used to compute inspiral, but may be changed by passing the keyword argument.

See also inertial_waveform for the waveform in the inertial frame.

Tip

If you need this waveform at a different set of times t′ than is currently present in inspiral.t, you should use the built-in interpolation capabilities of inspiral first, as in inspiral′ = inspiral(t′), rather than interpolating the results of this function. Or, perhaps better yet, you could select the times when calling orbital_evolution by using the saveat or saves_per_orbit keyword argument to that function. These approaches will be more accurate, faster, and require less memory than interpolating the result of this function.

The mode weights are given starting at ℓₘᵢₙ (which must satisfy 0 ≤ ℓₘᵢₙ ≤ 2) and extending through ℓₘₐₓ. The waveform is returned as a 2-dimensional Array, in which the first index varies over the mode index from (ℓ, m) = (ℓₘᵢₙ, -ℓₘᵢₙ) to (ℓ, m) = (ℓₘₐₓ, ℓₘₐₓ), with m varying most rapidly, and the second index varying over the time steps.

By default, the waveform is returned in the co-orbital frame — which is somewhat like the co-rotating frame. In particular, the modes of non-precessing systems vary slowly, over inspiral timescales; modes of precessing systems still vary on orbital timescales, though even this variation could be factored out. If inertial=true is passed, the waveform is instead transformed to the inertial frame, resulting in the oscillatory behavior we usually expect from a waveform.

PostNewtonian.inertial_waveform — Function

inertial_waveform(inspiral, [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Evaluate the post-Newtonian waveform mode weights in the inertial frame for the given inspiral output by orbital_evolution.

The inertial frame is the one in which inertial observers are found, so this waveform is more like one that actual observers would detect. This function transforms the waveform from the co-orbital frame — which is the one in which PN expressions are provided.

See coorbital_waveform for details about the other arguments.

In-place evaluation of waveforms

This is likely to be an uncommon scenario, but if you happen to need to evaluate the waveform repeatedly on the same set (or at least same number) of time steps and same range of $\ell$ values, then you can pre-allocate the storage needed in this process, and re-use that storage. To do so, pre-allocate using either coorbital_waveform_computation_storage or inertial_waveform_computation_storage, and then pass the result as the first argument to either coorbital_waveform! or inertial_waveform!.

Note

The h array returned by either coorbital_waveform! or inertial_waveform! is part of the pre-allocated storage. You need to use its values before you call either of those functions again, or those values will just be overwritten.

PostNewtonian.coorbital_waveform_computation_storage — Function

coorbital_waveform_computation_storage(inspiral; [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Construct storage needed to compute waveforms in the co-orbital frame without allocations.

This returns the storage for the waveforms themselves and PNSystem used as temporary storage. The returned quantity can just be passed as the first argument to coorbital_waveform! without being unpacked.

The meaning of the arguments is the same as in coorbital_waveform.

PostNewtonian.inertial_waveform_computation_storage — Function

inertial_waveform_computation_storage(inspiral; [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Construct storage needed to compute waveforms in the inertial frame without allocations.

This returns the storage for the waveforms themselves, storage used for computing the Wigner 𝔇 matrices, and for "in-place" multiplication. The returned quantity can just be passed as the first argument to inertial_waveform! without being unpacked.

The meaning of the arguments is the same as in inertial_waveform.

PostNewtonian.coorbital_waveform! — Function

coorbital_waveform!(storage, inspiral; [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Evaluate the post-Newtonian waveform mode weights in the co-orbital frame for the given inspiral output by orbital_evolution, using pre-allocated storage.

The storage is assumed to be the object returned from coorbital_waveform_computation_storage. Other arguments are the same as in coorbital_waveform.

PostNewtonian.inertial_waveform! — Function

inertial_waveform!(storage, inspiral; [ℓₘᵢₙ=2], [ℓₘₐₓ=8], [PNOrder])

Evaluate the post-Newtonian waveform mode weights in the inertial frame for the given inspiral output by orbital_evolution, using pre-allocated storage.

The storage is assumed to be the object returned from inertial_waveform_computation_storage. Other arguments are the same as in inertial_waveform.