Termination criteria for inspiral evolution

These are the termination criteria used by default in the inspiral function.

PostNewtonian.termination_forwards — Function

termination_forwards(vₑ, [quiet])

Construct termination criteria of solving PN evolution forwards in time

These criteria include checking that the masses are positive and the dimensionless spins are less than 1, as well as ensuring that the evolution will terminate at vₑ.

The optional quiet argument will silence informational messages about reaching the target value of vₑ if set to true, but warnings will still be issued when terminating for other reasons. If you want to quiet warnings also, you can do something like this:

using Logging
with_logger(SimpleLogger(Logging.Error)) do
    <your code goes here>
end

PostNewtonian.termination_backwards — Function

termination_backwards(v₁, [quiet])

Construct termination criteria of solving PN evolution backwards in time

These criteria include checking that the masses are positive and the dimensionless spins are less than 1, as well as ensuring that the evolution will terminate at v₁.

The optional quiet argument will silence informational messages about reaching the target value of v₁ if set to true, but warnings will still be issued when terminating for other reasons. If you want to quiet warnings also, you can do something like this:

using Logging
with_logger(SimpleLogger(Logging.Error)) do
    <your code goes here>
end

PostNewtonian.dtmin_terminator — Function

dtmin_terminator(T)

Construct termination criterion to terminate when dt drops below √eps(T).

Pass force_dtmin=true to solve when using this callback. Otherwise, the time-step size may decrease too much within a single time step, so that the integrator itself will quit before reaching this callback, leading to a less graceful exit.

PostNewtonian.nonfinite_terminator — Function

nonfinite_terminator()

Construct termination criterion to terminate when any NaN or Inf is found in the data after an integration step.