Mass-parameter conversions

PostNewtonian.q — Method

q(M₁, M₂)
mass_ratio(M₁, M₂)

Compute mass ratio $M₁/M₂$.

Note that we do not restrict to M₁ ≥ M₂ or vice versa; if you prefer that $q$ always be greater than or equal to 1 (or vice versa), you are responsible for ensuring that

PostNewtonian.δ — Method

δ(M₁, M₂)
mass_difference_ratio(M₁, M₂)

Compute mass-difference ratio $(M₁-M₂)/(M₁+M₂)$.

PostNewtonian.μ — Method

μ(M₁, M₂)
reduced_mass(M₁, M₂)

Compute the reduced mass $(M₁ M₂)/(M₁+M₂)$.

PostNewtonian.ν — Method

ν(M₁, M₂)
reduced_mass_ratio(M₁, M₂)

Compute the reduced mass ratio $(M₁ M₂)/(M₁+M₂)^2$.

Note that the denominator is squared, unlike in the reduced mass μ.

PostNewtonian.ℳ — Method

ℳ(M₁, M₂)
chirp_mass(M₁, M₂)

Compute the chirp mass ℳ, which determines the leading-order orbital evolution of a binary system due to energy loss by gravitational-wave emission.

The chirp mass is defined as

\[ \mathcal{M} = rac{(M₁ M₂)^{3/5}} {(M₁ + M₂)^{1/5}}.\]

Spin-parameter conversions

PostNewtonian.S — Method

S(M₁, M₂, χ⃗₁, χ⃗₂)

Total spin vector $S⃗₁+S⃗₂$.

PostNewtonian.Σ — Method

Σ(M₁, M₂, χ⃗₁, χ⃗₂)

Differential spin vector $M(a⃗₂-a⃗₁)$.

PostNewtonian.χₐ — Method

χₐ(M₁, M₂, χ⃗₁, χ⃗₂)

Antisymmetric spin vector (χ⃗₁-χ⃗₂)/2.

PostNewtonian.χₛ — Method

χₛ(M₁, M₂, χ⃗₁, χ⃗₂)

Symmetric spin vector (χ⃗₁+χ⃗₂)/2.

PostNewtonian.χ⃗ — Method

χ⃗(S⃗, M)

Normalized spin vector $S⃗/M²$.

Orbital elements

PostNewtonian.n̂ — Method

n̂(R)

The unit vector pointing from object 2 to object 1.

PostNewtonian.λ̂ — Method

λ̂(R)

The unit vector pointing in the direction of the instantaneous velocity of object 1. This also completes the right-handed triple of (ℓ̂, n̂, λ̂).

PostNewtonian.ℓ̂ — Method

ℓ̂(R)

The unit vector pointing along the direction of orbital angular velocity.

ODE solutions

PostNewtonian.combine_solutions — Method

combine_solutions(sol₋, sol₊)

Combine ODESolutions

This function is internal to this package. It is not entirely general, but allows us to combine the backwards- and forwards-in-time solutions of the PN inspiral ODE equations into a single ODESolution object that should behave just as if it were the result of solve. In particular, indexing, interpolation, and iterations should behave exactly as described in the DifferentialEquations docs.

Irrational constants

PostNewtonian.apery — Constant

ζ3
apery

Apéry's constant is defined as $ζ(3)$, where $ζ$ is the Riemann zeta function. It is OEIS sequence A002117.

PostNewtonian.ζ3 — Constant

ζ3
apery

Apéry's constant is defined as $ζ(3)$, where $ζ$ is the Riemann zeta function. It is OEIS sequence A002117.