Derived variables

Orbital elements

PostNewtonian.DerivedVariables.n̂ — Function

n̂(pnsystem)
n̂(R)
n_hat(pnsystem)
n_hat(R)

The unit vector pointing from object 2 to object 1, when the frame is given by the rotor R. This is equal to

\[n̂(R) = R x̂ R̄\]

PostNewtonian.DerivedVariables.λ̂ — Function

λ̂(pnsystem)
λ̂(R)
lambda_hat(pnsystem)
lambda_hat(R)

The unit vector pointing in the direction of the instantaneous velocity of object 1, when the frame is given by the rotor R. This is equal to

\[λ̂(R) = R ŷ R̄\]

This also completes the right-handed triple of $(n̂, λ̂, ℓ̂)$.

PostNewtonian.DerivedVariables.ℓ̂ — Function

ℓ̂(pnsystem)
ℓ̂(R)
ell_hat(pnsystem)
ell_hat(R)

The unit vector pointing along the direction of orbital angular velocity, when the frame is given by the rotor R. This is equal to

\[ℓ̂(R) = R ẑ R̄\]

PostNewtonian.DerivedVariables.Ω — Function

Ω(pnsystem)
Ω(;v, M=1)
Omega(pnsystem)
Omega(;v, M=1)

Orbital angular frequency.

The parameter v is the PN velocity parameter, and must be passed as a keyword argument — as in Ω(v=0.1). The parameter M is the total mass of the binary. They are related by definition as

\[\Omega \colonequals \frac{v^3}{M}.\]

See also v.

Mass combinations

PostNewtonian.DerivedVariables.M — Function

M(pnsystem)
M(M₁, M₂)
total_mass(pnsystem)
total_mass(M1, M2)

Compute the total mass $M₁+M₂$.

PostNewtonian.DerivedVariables.μ — Function

μ(pnsystem)
μ(M₁, M₂)
reduced_mass(pnsystem)
reduced_mass(M1, M2)

Compute the reduced mass $(M₁ M₂)/(M₁+M₂)$.

PostNewtonian.DerivedVariables.ν — Function

ν(pnsystem)
ν(M₁, M₂)
reduced_mass_ratio(pnsystem)
reduced_mass_ratio(M1, M2)

Compute the reduced mass ratio $(M₁ M₂)/(M₁+M₂)^2$.

Note that the denominator is squared, unlike in the reduced mass μ.

PostNewtonian.DerivedVariables.δ — Function

δ(pnsystem)
δ(M₁, M₂)
mass_difference_ratio(pnsystem)
mass_difference_ratio(M1, M2)

Compute mass-difference ratio $(M₁-M₂)/(M₁+M₂)$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $δ$ always be positive (or always negative), you are responsible for ensuring that.

PostNewtonian.DerivedVariables.q — Function

q(pnsystem)
q(M₁, M₂)
mass_ratio(pnsystem)
mass_ratio(M1, M2)

Compute mass ratio $M₁/M₂$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $q$ always be greater than or equal to 1 (or vice versa), you are responsible for ensuring that.

PostNewtonian.DerivedVariables.ℳ — Function

ℳ(pnsystem)
ℳ(M₁, M₂)
chirp_mass(pnsystem)
chirp_mass(M1, M2)

Compute the chirp mass ℳ, which determines the leading-order orbital evolution of a binary system due to energy loss by gravitational-wave emission.

The chirp mass is defined as

\[ \mathcal{M} = \frac{(M_1 M_2)^{3/5}} {(M_1 + M_2)^{1/5}}.\]

PostNewtonian.DerivedVariables.X₁ — Function

X₁(pnsystem)
X₁(M₁, M₂)
X1(pnsystem)
X1(M1, M2)

Compute the reduced individual mass $M₁/(M₁+M₂)$.

PostNewtonian.DerivedVariables.X₂ — Function

X₂(pnsystem)
X₂(M₁, M₂)
X2(pnsystem)
X2(M1, M2)

Compute the reduced individual mass $M₂/(M₁+M₂)$.

Spin combinations

PostNewtonian.DerivedVariables.S⃗₁ — Function

S⃗₁(pnsystem)

Dimensionful spin vector of object 1.

PostNewtonian.DerivedVariables.S⃗₂ — Function

S⃗₂(pnsystem)

Dimensionful spin vector of object 2.

PostNewtonian.DerivedVariables.S⃗ — Function

S⃗(pnsystem)
S⃗(M₁, M₂, χ⃗₁, χ⃗₂)

Total (dimensionful) spin vector $S⃗₁+S⃗₂$.

PostNewtonian.DerivedVariables.Σ⃗ — Function

Σ⃗(pnsystem)
Σ⃗(M₁, M₂, χ⃗₁, χ⃗₂)

Differential spin vector $M(a⃗₂-a⃗₁)$.

PostNewtonian.DerivedVariables.χ⃗ — Function

χ⃗(pnsystem)
χ⃗(S⃗, M)

Normalized spin vector $S⃗/M²$.

PostNewtonian.DerivedVariables.χ⃗ₛ — Function

χ⃗ₛ(M₁, M₂, χ⃗₁, χ⃗₂)

Symmetric spin vector $(χ⃗₁+χ⃗₂)/2$.

PostNewtonian.DerivedVariables.χ⃗ₐ — Function

χ⃗ₐ(M₁, M₂, χ⃗₁, χ⃗₂)

Antisymmetric spin vector $(χ⃗₁-χ⃗₂)/2$.

PostNewtonian.DerivedVariables.χₑ — Function

χₑ(s)
chi_eff(s)

Effective spin parameter of the system.

Defined as

\[\chi_{\mathrm{eff}} \colonequals \frac{c}{G} \left( \frac{\mathbf{S}_1}{M_1} + \frac{\mathbf{S}_2}{M_2} \right) \cdot \frac{\hat{\mathbf{L}}_{\mathrm{N}}} {M}.\]

PostNewtonian.DerivedVariables.χₚ — Function

χₚ(s)
chi_p(s)

Effective precession spin parameter of the system.

Note that there are two different definitions of this quantity found in the literature. The original definition (converted to the convention where $M_1 \geq M_2$) is

\[\begin{gathered} A_1 = 2 + \frac{3M_2}{2M_1} \\ A_2 = 2 + \frac{3M_1}{2M_2} \\ \chi_{\mathrm{p}} \colonequals \frac{1}{A_1 M_1^2} \mathrm{max}\left(A_1 S_{1\perp}, A_2 S_{2\perp} \right). \end{gathered}\]

In a paper from early in the detection era, the LIGO collaboration used this definition.

However, a more recent paper redefines this essentially as $M_1^2$ times that quantity. Using the convention that $q = M_2/M_1 \leq 1$, the definition may be more compactly written as

\[\chi_{\mathrm{p}} \colonequals \mathrm{max} \left( \chi_{1\perp}, \chi_{2\perp} q \frac{4q+3}{4+3q} \right).\]

Again, a more recent paper by LIGO/Virgo/KAGRA uses this convention.

Because it seems to be the trend, this function uses the latter definition.

PostNewtonian.DerivedVariables.S⃗₀⁺ — Function

S⃗₀⁺(s)
S⃗₀⁺(M₁, M₂, κ₁, κ₂, S⃗₁, S⃗₂)

Defined in Eq. (3.4) of Buonanno et al. (2012):

\[\vec{S}_0^+ = \frac{M}{M_1} \left( \frac{\kappa_1} {\kappa_2} \right)^{1/4} \left( 1 + \sqrt{1 - \kappa_1 \kappa_2} \right)^{1/2} \vec{S}_1 + \frac{M}{M_2} \left( \frac{\kappa_2} {\kappa_1} \right)^{1/4} \left( 1 - \sqrt{1 - \kappa_1 \kappa_2} \right)^{1/2} \vec{S}_2.\]

Note that, currently, $\kappa_1$ and $\kappa_2$ are both assumed to be equal to 1, as is the case for black holes. You can define κ₁ and κ₂ to have other values for your own PNSystem types, and this function will work appropriately.

See also S⃗₀⁻.

PostNewtonian.DerivedVariables.S⃗₀⁻ — Function

S⃗₀⁻(s)
S⃗₀⁻(M₁, M₂, κ₁, κ₂, S⃗₁, S⃗₂)

Defined below Eq. (3.4) of Buonanno et al. (2012):

\[\vec{S}_0^- = \frac{M}{M_1} \left( \frac{\kappa_1} {\kappa_2} \right)^{1/4} \left( 1 - \sqrt{1 - \kappa_1 \kappa_2} \right)^{1/2} \vec{S}_1 + \frac{M}{M_2} \left( \frac{\kappa_2} {\kappa_1} \right)^{1/4} \left( 1 + \sqrt{1 - \kappa_1 \kappa_2} \right)^{1/2} \vec{S}_2.\]

Note that, currently, $\kappa_1$ and $\kappa_2$ are both assumed to be equal to 1, as is the case for black holes. You can define κ₁ and κ₂ to have other values for your own PNSystem types, and this function will work appropriately.

See also S⃗₀⁺.

Additionally, there are numerous convenience functions to give certain components of the spins. They take a single pnsystem argument and are not exported. Given the definitions above, they are all fairly self explanatory — such as χ₁², which gives χ⃗₁ ⋅ χ⃗₁; or χ₁₂ = χ⃗₁ ⋅ χ⃗₂; or Sₙ = S⃗ ⋅ n̂. Like all the other fundamental and derived variables, these can be used directly in PN expressions modified by the @pn_expression macro.

Horizons

We can also compute some variables defined by Alvi (2001) related to the horizons. The hardest parts to compute here involve the relative angles between the spins and the black-hole separation vectors. Alvi constructs a spherical coordinate system centered on each black hole where the $z$ axis is given by the direction of the spin, and $\theta$ and $\phi$ represent the direction to the other black hole. While he makes a (somewhat ambiguous) choice about the origin of the $\phi$ coordinate, only $\dot{\phi}$ comes into the equations, so we don't really care about that origin.

Note that Alvi uses $\mathbf{n}$ to represent the "normal to the orbital plane", whereas we — and most of the rest of the post-Newtonian literature — use $\hat{\ell}$ for this vector and $\hat{n}$ to represent the separation vector pointing from object 2 to object 1. For convenience, we define

\[\hat{n}_i = \begin{cases} -\hat{n} & \text{i=1}, \\ \hphantom{-}\hat{n} & \text{i=2}. \end{cases}\]

Alvi's construction is also somewhat adiabatic, so we treat the spins and orbital plane as constant in the calculation of the instantaneous tidal heating — though they evolve slowly over time — and the angles $\theta$ and $\phi$ as evolving rapidly. In our formulation, then, the goal is to find the angle $\theta_i(t)$ between $\vec{\chi}_i$ and $\hat{n}_i$, and the rotation rate $\dot{\phi}_i(t)$ of $\hat{n}_i$ about the $\vec{\chi}_i$ axis.

Computing the angle between vectors is a somewhat infamously tricky problem. There are various claims floating around about the best ways to compute quantities involving areas and angles of triangles. While these claims are surely true for areas, I am more skeptical of the relevance for angles. I find it best to realize that you probably don't need the angle per se, but trigonometric functions of the angle — like $\sin^2 \theta_i$, which is what we actually need in this case. In particular, I believe the best results come from computing

\[\sin^2\theta_i = \frac{ \left|\hat{n}_i \times \vec{\chi}_i\right|^2 }{ \left| \vec{\chi}_i\right|^2 }.\]

If the denominator is zero, we set $\sin^2\theta_i = 1$ from physical considerations.

Now consider the quantity $\hat{n}_i \times \vec{\chi}_i$. We next aim to calculate the rotation rate $\dot{\phi}_i$ of this vector about $\vec{\chi}_i$. We begin by directly calculating

\[\partial_t \left(\hat{n}_i \times \vec{\chi}_i\right) = \left(\partial_t \hat{n}_i\right) \times \vec{\chi}_i = \left(\Omega\, \hat{\ell} \times \hat{n}_i\right) \times \vec{\chi}_i = \mp \Omega\, \hat{\lambda} \times \vec{\chi}_i,\]

where the negative sign is chosen for $i=1$ and positive for $i=2$. Now, from more fundamental considerations, we can understand the components of this change. Since we assume that $\vec{\chi}_i$ is constant at each instant for the purposes of calculation here, the only way for $\hat{n}_i \times \vec{\chi}_i$ to change is either because $\hat{n}_i$ rotates about $\vec{\chi}_i$, or because the angle $\theta_i$ is changing. We can express this as

\[\partial_t \left(\hat{n}_i \times \vec{\chi}_i\right) = \left( \dot{\phi}\, \hat{\chi}_i \right) \times \left(\hat{n}_i \times \vec{\chi}_i\right) + \dot{\theta_i}\, \cot \theta_i\, \left(\hat{n}_i \times \vec{\chi}_i\right).\]

Since these two components are orthogonal, we can obtain $\dot{\phi}$ directly by taking the component of this quantity along $\hat{\chi}_i \times \left(\hat{n}_i \times \vec{\chi}_i\right)$:

\[\dot{\phi}_i = \frac{ \left( \Omega\, \hat{\lambda} \times \vec{\chi}_i \right) \cdot \left( \vphantom{\hat{\lambda}} \hat{\chi}_i \times \left(\hat{n} \times \vec{\chi}_i\right) \right) }{ \left| \hat{\chi}_i \times \left(\hat{n} \times \vec{\chi}_i\right) \right|^2 } = \Omega\, \frac{\hat{\ell} \cdot \hat{\chi}_i}{\sin^2 \theta_i}.\]

Here again, we may run into numerical trouble if $\left| \vec{\chi}_1 \right| \approx 0$, in which case we again use physical arguments to take $\dot{\phi}_i = \Omega$. We might also expect to run into trouble if $\sin^2 \theta_i \approx 0$, which corresponds to a polar orbit, in which case Alvi's approximations break down. This turns out to not be a problem numerically, because of the cancellation with the numerator, except when $\sin^2 \theta_i = 0$ exactly. In this case, we choose $\dot{\phi}_i = 0$.

Note that the sign of $\hat{n}_i$ has dropped out of the calculations of both $\sin^2\theta_i$ and $\dot{\phi}_i$, cancelling with the signs that had appeared next to $\Omega$.

PostNewtonian.DerivedVariables.rₕ₁ — Function

rₕ₁(s)

Horizon radius of black hole 1.

As defined on page 2, line 4, of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.rₕ₂ — Function

rₕ₂(s)

Horizon radius of black hole 2.

As defined on page 2, line 4, of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.Ωₕ₁ — Function

Ωₕ₁(s)

Horizon angular velocity of black hole 1.

As defined on page 2, line 5, of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.Ωₕ₂ — Function

Ωₕ₂(s)

Horizon angular velocity of black hole 2.

As defined on page 2, line 5, of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.sin²θ₁ — Function

sin²θ₁(s)

Sine-squared of angle between spin of black hole 1 and vector to black hole 2.

Compare to Eq. (18) of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.sin²θ₂ — Function

sin²θ₂(s)

Sine-squared of angle between spin of black hole 2 and vector to black hole 1.

Compare to Eq. (18) of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.ϕ̇̂₁ — Function

ϕ̇̂₁(s)

Rate of rotation of black hole 2 about the spin of black hole 1, relative to orbital rotation rate.

This is the rotation rate ϕ̇ as defined in Eq. (19) of Alvi (2001), divided by $v^3 = M\, \Omega$. This division is done to make sure we can track the relative PN order of terms that depend on this.

See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.ϕ̇̂₂ — Function

ϕ̇̂₂(s)

Rate of rotation of black hole 1 about the spin of black hole 2, relative to orbital rotation rate.

This is the rotation rate ϕ̇ as defined in Eq. (19) of Alvi (2001), divided by $v^3 = M\, \Omega$. This division is done to make sure we can track the relative PN order of terms that depend on this.

See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.Î₀₁ — Function

Î₀₁(s)

Horizon moment of inertia of black hole 1.

This is the moment divided by $ν^2 v^{12}$, as given by Eq. (10) of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.Î₀₂ — Function

Î₀₂(s)

Horizon moment of inertia of black hole 2.

This is the moment divided by $ν^2 v^{12}$, as given by Eq. (10) of Alvi (2001). See the documentation section on "Horizons" for more details.

PostNewtonian.DerivedVariables.κ₁ — Function

κ₁(s)

The "quadrupolar polarisability" of object 1 used by Bohé et al. (2015).

Note that Bohé et al. refer to the closely related (and co-authored) Marsat (2014), who notes above Eq. (4.7) that this is denoted $C_{\mathrm{ES}^2}$ in Levi and Steinhoff (2014), who in turn notes that "we can set ... the Wilson coefficients $C_{\mathrm{ES}^2} = C_{\mathrm{BS}^3} = 1$ for the black hole case."

However, a very similar constant $\kappa_A$ is used in Eq. (2.1) of Buonanno et al. (2012). They say that $\kappa_A=1$ for an isolated black hole, but can deviate from 1 for a black hole in a binary — though those deviations occur at "much higher" order than those they consider (2PN).

See also λ₁.

Warn

This function will be incorrect for objects other than black holes. It is not clear to me if this is the same quantity as $C_Q$ used in some papers, such as Bini and Geralico (2014), but they point out that for neutron stars, the value varies between 4.3 and 7.4, depending on the equation of state. This quantity may also be related to λ₁. Pull requests or issues with more information are welcome.

PostNewtonian.DerivedVariables.κ₂ — Function

κ₂(s)

The "quadrupolar polarisability" of object 2 used by Bohé et al. (2015). See κ₁ for more details.

PostNewtonian.DerivedVariables.κ₊ — Function

κ₊(s)

Equal to κ₁+κ₂; defined below Eq. (3.28) of Bohé et al. (2015).

PostNewtonian.DerivedVariables.κ₋ — Function

κ₋(s)

Equal to κ₁-κ₂; defined below Eq. (3.28) of Bohé et al. (2015).

PostNewtonian.DerivedVariables.λ₁ — Function

λ₁(s)

The "quadrupolar polarisability" of object 1 used by Marsat (2014), who notes above Eq. (4.11) that this is denoted $C_{\mathrm{BS}^2}$ in Levi and Steinhoff (2014), who in turn notes that "we can set ... the Wilson coefficients $C_{\mathrm{ES}^2} = C_{\mathrm{BS}^3} = 1$ for the black hole case."

See also κ₁.

Warn

This function will be incorrect for objects other than black holes. It is not clear to me if this is the same quantity as $C_Q$ used in some papers, such as Bini and Geralico (2014), but they point out that for neutron stars, the value varies between 4.3 and 7.4, depending on the equation of state. This quantity may also be related to λ₁. Pull requests or issues with more information are welcome.

PostNewtonian.DerivedVariables.λ₂ — Function

λ₂(s)

The "quadrupolar polarisability" of object 2 used by Bohé et al. (2015). See λ₁ for more details.

PostNewtonian.DerivedVariables.λ₊ — Function

λ₊(s)

Equal to λ₁+λ₂; defined below Eq. (3.28) of Bohé et al. (2015).

PostNewtonian.DerivedVariables.λ₋ — Function

λ₋(s)

Equal to λ₁-λ₂; defined below Eq. (3.28) of Bohé et al. (2015).