Fundamental variables

These constitute the basic quantities needed to fully describe a PN binary system at any instant. All other variables — and ultimately all PN expressions — may be expressed as functions of these. In the docstrings below, pnsystem is an object of any type that subtypes PNSystem.

PostNewtonian.FundamentalVariables.M₁ — Function

M₁(pnsystem)
M1(pnsystem)

Mass of object 1 in this system.

PostNewtonian.FundamentalVariables.M₂ — Function

M₂(pnsystem)
M2(pnsystem)

Mass of object 2 in this system.

PostNewtonian.FundamentalVariables.χ⃗₁ — Function

χ⃗₁(pnsystem)
chi1(pnsystem)

Dimensionless spin vector of object 1 in this system, as a QuatVec.

PostNewtonian.FundamentalVariables.χ⃗₂ — Function

χ⃗₂(pnsystem)
chi2(pnsystem)

Dimensionless spin vector of object 2 in this system, as a QuatVec.

PostNewtonian.FundamentalVariables.R — Function

R(pnsystem)

Orientation of the binary, as a Rotor. This Rotor will rotate the ẑ vector to be along the orbital angular velocity vector and the x̂ vector to be along the separation vector pointing from object 2 to object 1.

See also n̂, λ̂, ℓ̂.

PostNewtonian.FundamentalVariables.v — Function

v(pnsystem)
v(;Ω, M=1)

Post-Newtonian velocity parameter. This is related to the orbital angular frequency $\Omega$ as

\[v \colonequals (M\,\Omega)^{1/3},\]

where $M$ is the total mass of the binary.

Note that if you want to pass the value $Ω$ (rather than a PNSystem), you must pass it as a keyword argument — as in v(Ω=0.1).

See also Ω.

PostNewtonian.FundamentalVariables.Φ — Function

Φ(pnsystem)
Phi(pnsystem)

Integrated orbital phase of the system. It is computed as the integral of Ω.

PostNewtonian.FundamentalVariables.λ₁ — Function

λ₁(pnsystem)
lambda1(pnsystem)

Tidal coupling parameter of object 1 in this system.

This is related to the Love number $k_2$ and the star's radius $R$ as

\[\lambda_1 = \frac{2}{3} k_2 R^5.\]

There is another common parameter denoted

\[\Lambda_1 = \frac{3}{2} \frac{M_1}{M_2} \lambda_1.\]

Note that — as of this writing — only NSNS systems can have a nonzero value for this quantity. All other types return 0, which Julia can use to eliminate code that would then be 0. Thus, it is safe and efficient to use this quantity in any PN expression that specializes on the type of pnsystem.

PostNewtonian.FundamentalVariables.λ₂ — Function

λ₂(pnsystem)
lambda2(pnsystem)

Tidal coupling parameter of object 2 in this system.

This is related to the Love number $k_2$ and the star's radius $R$ as

\[\lambda_2 = \frac{2}{3} k_2 R^5.\]

There is another common parameter denoted

\[\Lambda_2 = \frac{3}{2} \frac{M_2}{M_1} \lambda_2.\]

Note that — as of this writing — only BHNS and NSNS systems can have a nonzero value for this quantity. All other types return 0, which Julia can use to eliminate code that would then be 0. Thus, it is safe and efficient to use this quantity in any PN expression that specializes on the type of pnsystem.