Mass-parameter conversions

PostNewtonian.q — Method

q(M₁, M₂)
mass_ratio(M₁, M₂)

Compute mass ratio $M₁/M₂$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $q$ always be greater than or equal to 1 (or vice versa), you are responsible for ensuring that.

PostNewtonian.δ — Method

δ(M₁, M₂)
mass_difference_ratio(M₁, M₂)

Compute mass-difference ratio $(M₁-M₂)/(M₁+M₂)$.

Note that we do not restrict to $M₁ ≥ M₂$ or vice versa; if you prefer that $δ$ always be positive (or always negative), you are responsible for ensuring that.

PostNewtonian.μ — Method

μ(M₁, M₂)
reduced_mass(M₁, M₂)

Compute the reduced mass $(M₁ M₂)/(M₁+M₂)$.

PostNewtonian.ν — Method

ν(M₁, M₂)
reduced_mass_ratio(M₁, M₂)

Compute the reduced mass ratio $(M₁ M₂)/(M₁+M₂)^2$.

Note that the denominator is squared, unlike in the reduced mass μ.

PostNewtonian.ℳ — Method

ℳ(M₁, M₂)
chirp_mass(M₁, M₂)

Compute the chirp mass ℳ, which determines the leading-order orbital evolution of a binary system due to energy loss by gravitational-wave emission.

The chirp mass is defined as

\[ \mathcal{M} = \frac{(M₁ M₂)^{3/5}} {(M₁ + M₂)^{1/5}}.\]

Spin-parameter conversions

PostNewtonian.S — Method

S(M₁, M₂, χ⃗₁, χ⃗₂)

Total spin vector $S⃗₁+S⃗₂$.

PostNewtonian.Σ — Method

Σ(M₁, M₂, χ⃗₁, χ⃗₂)

Differential spin vector $M(a⃗₂-a⃗₁)$.

PostNewtonian.χₐ — Method

χₐ(M₁, M₂, χ⃗₁, χ⃗₂)

Antisymmetric spin vector $(χ⃗₁-χ⃗₂)/2$.

PostNewtonian.χₛ — Method

χₛ(M₁, M₂, χ⃗₁, χ⃗₂)

Symmetric spin vector $(χ⃗₁+χ⃗₂)/2$.

PostNewtonian.χ⃗ — Method

χ⃗(S⃗, M)

Normalized spin vector $S⃗/M²$.

Orbital elements

PostNewtonian.n̂ — Function

n̂(R)
n_hat(R)

The unit vector pointing from object 2 to object 1, when the frame is given by the rotor R. This is equal to

\[n̂(R) = R x̂ R̄\]

PostNewtonian.λ̂ — Function

λ̂(R)
lambda_hat(R)

The unit vector pointing in the direction of the instantaneous velocity of object 1, when the frame is given by the rotor R. This is equal to

\[λ̂(R) = R ŷ R̄\]

This also completes the right-handed triple of $(ℓ̂, n̂, λ̂)$.

PostNewtonian.ℓ̂ — Function

ℓ̂(R)
ell_hat(R)

The unit vector pointing along the direction of orbital angular velocity, when the frame is given by the rotor R. This is equal to

\[ℓ̂(R) = R ẑ R̄\]

PostNewtonian.v — Function

v(;Ω, M=1)

Post-Newtonian velocity parameter.

The parameter Ω is the orbital angular frequency, and must be passed as a keyword argument — as in v(Ω=0.1). The parameter M is the total mass of the binary.

PostNewtonian.Ω — Function

Ω(;v, M=1)
Omega(;v, M=1)

Orbital angular frequency.

The parameter v is the PN velocity parameter, and must be passed as a keyword argument — as in Ω(v=0.1). The parameter M is the total mass of the binary.

ODE solutions

PostNewtonian.combine_solutions — Method

combine_solutions(sol₋, sol₊)

Combine ODESolutions

This function is internal to this package. It is not entirely general, but allows us to combine the backwards- and forwards-in-time solutions of the PN inspiral ODE equations into a single ODESolution object that should behave just as if it were the result of solve. In particular, indexing, interpolation, and iterations should behave exactly as described in the DifferentialEquations docs.

Irrational constants

These quantities are constants that appear in PN expressions, so they are not exported, but can be used by importing them explicitly or by using the fully qualified names. They are defined here as Irrationals. This means that Julia can convert them to float types as necessary. Unfortunately, by default Julia converts to Float64. For example, BigFloat(2ζ3) will be a BigFloat, but will only have the precision of a Float64, because 2ζ3 is converted first. To get full precision, you'll need to do things like 2BigFloat(ζ3).

PostNewtonian.log2 — Constant

log2

The natural logarithm of 2. This is OEIS sequence A002162.

julia> PostNewtonian.log2
log2 = 0.6931471805599...

julia> exp(PostNewtonian.log2)
2.0

julia> exp(big(PostNewtonian.log2))
2.0

PostNewtonian.log3 — Constant

log3

The natural logarithm of 3. This is OEIS sequence A002391.

julia> PostNewtonian.log3
log3 = 1.0986122886681...

julia> exp(PostNewtonian.log3)
3.0000000000000004

julia> exp(big(PostNewtonian.log3))
3.0

PostNewtonian.log3halves — Constant

log3halves

The natural logarithm of 3//2. This is OEIS sequence A016578.

julia> PostNewtonian.log3halves
log3halves = 0.4054651081081...

julia> exp(PostNewtonian.log3halves)
1.5

julia> exp(big(PostNewtonian.log3halves))
1.5

PostNewtonian.log5 — Constant

log5

The natural logarithm of 5. This is OEIS sequence A016628.

julia> PostNewtonian.log5
log5 = 1.6094379124341...

julia> exp(PostNewtonian.log5)
4.999999999999999

julia> exp(big(PostNewtonian.log5))
5.0

PostNewtonian.log5halves — Constant

log5halves

The natural logarithm of 5//2. This is OEIS sequence A016579.

julia> PostNewtonian.log5halves
log5halves = 0.9162907318741...

julia> exp(PostNewtonian.log5halves)
2.5

julia> exp(big(PostNewtonian.log5halves))
2.5

PostNewtonian.ζ3 — Constant

ζ3
apery

Apéry's constant is defined as $ζ(3)$, where $ζ$ is the Riemann zeta function. This is OEIS sequence A002117.

julia> PostNewtonian.apery
apery = 1.2020569031595...

julia> PostNewtonian.ζ3
apery = 1.2020569031595...

julia> sum((1:10_000_000).^-3) - PostNewtonian.apery
-4.6629367034256575e-15